Fix test helps.

+++ tests/helpers/nonce.py
@@ -7,7 +7,7 @@ from ocean_provider.user_nonce import get_nonce, update_nonce
 def build_nonce(address) -> Decimal:
     nonce = get_nonce(address)
     if nonce:
-        nonce = Decimal(nonce) + 1
+        nonce = int(float(nonce)) + 1
         update_nonce(address, nonce)
 
         return Decimal(nonce)
